Refactor directory structure
Enforce module structure by creating a new `common/` directory and putting all the `util/` files in it along with AppTypeAhead.jsx and SingleSearchSelect.jsx components
